move compile-time magic to separate header
